Collection size (#32)
* Add collection size column

- add evaluationSize column to article
- add commentSize column to evaluation

* Change return value for vote pressing

- add total vote number to dto

+++ src/main/java/network/pluto/absolute/dto/MemberDto.java
@@ -45,7 +45,7 @@ public class MemberDto {
 
     private String major;
 
-    private int reputation;
+    private long reputation;
 
     @ApiModelProperty(readOnly = true)
     private WalletDto wallet;
